Handle buffer reload via :edit by making it dependent on b:changedtick as well
Rename the re-inclusion guard from b:SyntaxInclude_IncludedFiletypes (List) to b:SyntaxInclude_Included (Dict, with changedtick and filetypes keys). The purpose of the guard is avoid inclusion when multiple ranges are specified; this is probably done all at once in an after/syntax/*.vim script or via an :autocmd. Therefore, adding the current change number to the check still serves the purpose of avoid unnecessary re-inclusions, but will re-include when the buffer is reloaded (via :edit, which increments the change number) or otherwise re-initialized (:setlocal syntax=...) - cases where the buffer-local variable will persist.

Fixes #7